refactor file structure for witness chain
